🐛 fix: english inflection fixes (#340)
* 🐛 fix: add support to 'be' inflection

eventually will help when inflecting phrasal meanings such as 'be made' or 'be afraid'
adding test which helped find this bug

* 🐛 fix: fix incorrect penn pos tags in adjective inflection

* 🐛 fix: add adverb inflection using lemminflect

updating tests to match this

* 🚨 tests: updating integration test snapshots

* ✨ feat: add exclamation marks to imperative inflection

original inflection without exclamation marks remains

* 🚨 tests: fix broken tests

* 🔨 refactor: removing superfluous comments

* 🐛 fix: add support to 'can' inflection

This assumes that 'can' is never used for its other meaning, i.e. to preserve food in a can.
It also implements a dummy 'canning' for present participles. Unfortunately there is no alternative there.

* 🧹 chore: remove fixme resolved by 2957cc70